// data.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface Schema {
  meta: {
    [k: string]: unknown;
  };
  text: {
    id: string;
    name: string;
    path: string;
    base: string[];
    [k: string]: unknown;
  }[];
  books: {
    id: string;
    name: string;
    chapter: string[];
    [k: string]: unknown;
  }[];
  qa: {
    id: string;
    name: string;
    path: string;
    needAnswer?: boolean;
    [k: string]: unknown;
  }[];
  [k: string]: unknown;
}